Use a stats data structure for move times

+++ modules/api/src/main/java/net/countercraft/movecraft/craft/Craft.java
@@ -135,7 +135,7 @@ public interface Craft {
 
     long getOrigPilotTime();
 
-    float getMeanCruiseTime();
+    double getMeanCruiseTime();
 
     void addCruiseTime(float cruiseTime);
